working draft without good redirect's counting

+++ url_shortener/views.py
@@ -5,7 +5,6 @@ from url_shortener.forms import LinkShortenerForm
 from .models import Link
 
 
-
 def index(request):
     if request.method == 'POST':
         form = LinkShortenerForm(request.POST)
@@ -15,18 +14,13 @@ def index(request):
             return redirect('info/%s' % str(link.id))
     else:
         form = LinkShortenerForm()
-        top_url = Link.objects.all().order_by('-clicks_count')[:20]
+        top_url = Link.objects.all().order_by('-clicks_count', '-created')[:20]
     return render(request, 'url_shortener/index.html', {'form': form, 'top_url': top_url})
 
 
 def info(request, link_id):
     url = Link.objects.get(id=link_id)
     return render(request, 'url_shortener/info.html', {'url': url})
-
-
-# def overall(request):
-#     objects_list = Link.objects.all()
-#     return render(request, 'url_shortener/overall.html', {'objects_list': objects_list})
 
 
 def url_redirect(request, short):
@@ -37,7 +31,8 @@ def url_redirect(request, short):
 
 
 def delete_obj(request, link_id):
-    Link.objects.filter(id=link_id).delete()
+    if request.method == 'POST':
+        Link.objects.filter(id=link_id).delete()
     return redirect('/shortener/overall/')
